fix: load and network switch
