+++ apps/main/src/app/dex/layout.tsx
@@ -1,6 +1,8 @@
 import type { ReactNode } from 'react'
-import { App } from './client'
+import { App } from '@/app/dex/client'
 
+// We need to keep this layout here otherwise the whole app gets unmounted on network change
+// Because of that we cannot parse the URL params in the server component
 const Layout = ({ children }: { children: ReactNode }) => <App>{children}</App>
 
 export default Layout
