+++ packages/curve-ui-kit/src/utils/useTraceProps.ts
@@ -1,23 +1,6 @@
+'use client'
 import { useEffect, useRef } from 'react'
 
-/**
- * A hook used for debugging purposes to log the changes in the props of a component.
- * Even if left unused when committed, it's useful to leave it in the codebase for future debugging.
- * @see https://stackoverflow.com/a/51082563
- *
- * Usage:
- * ```tsx
- * import { useTraceProps } from '@ui-kit/utils/useTraceProps'
- *
- * const MyComponent = ({ prop1, prop2, prop3 }) => {
- *   useEffect(() => {
- *      // Your component logic
- *   }, [prop1, prop2])
- *   useTraceProps('MyComponent.useEffect', { prop1, prop2 })
- *   // ...
- * }
- */
-// noinspection JSUnusedGlobalSymbols
 export function useTraceProps<T extends Record<string, unknown>>(name: string, props: T) {
   const propsRef = useRef<T>(props)
   useEffect(() => {
@@ -35,10 +18,10 @@ export function useTraceProps<T extends Record<string, unknown>>(name: string, p
     }
     propsRef.current = props
   }, [props, name])
-  useEffect(
-    () => () => console.warn(`useTraceProps ${name} ${new Date().toISOString()}: Unmounted`, propsRef.current),
-    [name],
-  )
+  useEffect(() => {
+    console.info(`useTraceProps ${name} ${new Date().toISOString()}: Mounted`, propsRef.current)
+    return () => console.warn(`useTraceProps ${name} ${new Date().toISOString()}: Unmounted`, propsRef.current)
+  }, [name])
 }
 
 const toString = (value: unknown) => {
